requirements.txt
fastapi
uvicorn[standard]
email-validator
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
slowapi==0.1.9
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
